webapp/requirements.txt
flask
azure-storage-blob
requests
azure-cosmos
azure-mgmt-cosmosdb
azure-ai-formrecognizer
